- Downported to 1.19.2 - Added Mekanism Pipes - Added Refined Storage Cable

+++ src/main/java/com/heaser/pipeconnector/items/pipeconnectoritem/utils/ParticleHelper.java
@@ -1,10 +1,10 @@
 package com.heaser.pipeconnector.items.pipeconnectoritem.utils;
 
+import com.mojang.math.Vector3d;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.particles.DustParticleOptions;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.world.level.Level;
-import org.joml.Vector3f;
+
 
 public class ParticleHelper {
     public static void spawnDirectionHighlightParticles(Level level, BlockPos pos, float red, float green, float blue, float alpha) {
@@ -13,7 +13,7 @@ public class ParticleHelper {
         double z = pos.getZ() + 0.5;
 
         // Keeping this instead I'll want to use more complex particles
-        Vector3f color = new Vector3f(red, green, blue);
+        Vector3d color = new Vector3d(red, green, blue);
 
         for (int i = 0; i < 10; i++) {
             double offsetX = (Math.random() - 0.1) * 0.2;
